+++ src/algos/quality.c
@@ -174,7 +174,7 @@ static double QualityEstimate_ushort(fits *fit, int layer) {
 #endif
 		q = Gradient(buf, x_samples, y_samples);
 
-		dval += (q * ((QSUBSAMPLE_MIN * QSUBSAMPLE_MIN)
+		dval += (q * ((double)(QSUBSAMPLE_MIN * QSUBSAMPLE_MIN)
 					/ (subsample * subsample)));
 		//printf("dval val : %f\n", dval);
